cardpass-be/pyproject.toml
[project]
name = "cardpass-be"
version = "0.1.0"
description = "Web3 recruiting backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "base58>=2.1.1",
    "fastapi[standard]>=0.110.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pynacl>=1.5.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[depenency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]
